requirements.txt
Django==3.1.1
djangorestframework==3.11.1
flake8==3.8.3
psycopg2==2.8.6
Pillow>=5.3.0,<5.4.0
